feat: get and format schedule data
