+++ .gitignore
@@ -1,3 +1,2 @@
 /node_modules
 .env
-package-lock.json
