+++ button.js
@@ -1,23 +1,35 @@
-const monday = `
-Понедельник:
+const scheduleData = require('./data/events.js')
+// TODO: rewrite to ordinary function
+Date.prototype.getWeek = function () {
+  const target = new Date(this.valueOf());
+  const dayNr = (this.getDay() + 6) % 7;
+  target.setDate(target.getDate() - dayNr + 3);
+  const firstThursday = target.valueOf();
+  target.setMonth(0, 1);
+  if (target.getDay() !== 4) {
+    target.setMonth(0, 1 + ((4 - target.getDay() + 7) % 7));
+  }
+  return 1 + Math.ceil((firstThursday - target) / 604800000);
+};
 
-⏰ 9:55 — 11:25
-Технический перевод
-(ПЗ 2-14 н)
-ЛК - 6 этаж
+const isWeekEven = (date) => date.getWeek() % 2 === 0;
+const isBelongsToCurrentWeek = (date) => date.getWeek() === new Date().getWeek();
+// TODO: remove lessons from other years from datasets
+const isBelongsToCurrentYear = (date) => date.getFullYear() === new Date().getFullYear();
 
-⏰ 11:35 — 13:05
-Коммерческая оценка инвестиций
-(Л 2 - 12 н)
-Онищенко С.И.
-ЛК - 402
+const isCurrentWeekEven = isWeekEven(new Date());
 
-⏰ 13:45 — 15:15
-Технологический аудит
-(П3 2-14 н)
-Курочкин Д.А.
-ЛК - 402
-`;
+const mondayLessons = scheduleData.filter(({ start }) => isBelongsToCurrentYear(new Date(start)) && isBelongsToCurrentWeek(new Date(start)) && new Date(start).getDay() === 1)
+
+const formatLessons = (day, lessons ) => lessons.map(({ start, end, title, /* description */ }) =>`${day}\n⏰ ${new Date(start)} — ${new Date(end)}\n${title}\n`).join('\n\n');
+
+module.exports.schedule = {
+  monday: {
+    lessons: formatLessons('Понедельник', isCurrentWeekEven
+      ? mondayLessons.filter((lesson) => isWeekEven(new Date(lesson)) )
+      : mondayLessons.filter((lesson) => !isWeekEven(new Date(lesson))))
+  }
+}
 
 const tuesday = `
 Вторник:
@@ -287,7 +299,6 @@ const oddSaturday = `
 ЛК - 321
 `;
 
-module.exports.monday = monday;
 module.exports.tuesday = tuesday;
 module.exports.wensday = wensday;
 module.exports.thursday = thursday;
