+++ data/events.js
@@ -1,0 +1,1 @@
+module.exports = [{"id":763,"title":"Цифровая логистика инновационного проекта","color":"#0600FF","start":"2023-11-10T13:45:00","end":"2023-11-10T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК508<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Зубаков Геннадий Викторович<br><b>Кафедра:</b> Логистики"},{"id":3714,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-12-20T11:35:00","end":"2023-12-20T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":3711,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-12-21T09:55:00","end":"2023-12-21T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3712,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-12-21T08:15:00","end":"2023-12-21T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3713,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-12-20T13:45:00","end":"2023-12-20T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3715,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-12-14T09:55:00","end":"2023-12-14T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3717,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-12-07T09:55:00","end":"2023-12-07T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3719,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-30T09:55:00","end":"2023-11-30T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3721,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-23T09:55:00","end":"2023-11-23T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3723,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-16T09:55:00","end":"2023-11-16T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3725,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-09T09:55:00","end":"2023-11-09T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3784,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-02T09:55:00","end":"2023-11-02T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":3716,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-12-13T13:45:00","end":"2023-12-13T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":3718,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-12-06T13:45:00","end":"2023-12-06T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":3720,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-29T13:45:00","end":"2023-11-29T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":3722,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-22T13:45:00","end":"2023-11-22T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":3724,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-15T13:45:00","end":"2023-11-15T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":3783,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-08T13:45:00","end":"2023-11-08T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":3785,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-01T13:45:00","end":"2023-11-01T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":5713,"title":"Цифровая логистика инновационного проекта","color":"#0600FF","start":"2023-11-10T11:35:00","end":"2023-11-10T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК508<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Зубаков Геннадий Викторович<br><b>Кафедра:</b> Логистики"},{"id":25440,"title":"Коммерческая оценка инвестиций","color":"#0600FF","start":"2023-11-13T11:35:00","end":"2023-11-13T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Онищенко Светлана Ивановна<br><b>Кафедра:</b> Управления инновациями"},{"id":30999,"title":"Технологический аудит","color":"#0600FF","start":"2023-11-24T09:55:00","end":"2023-11-24T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК110<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Курочкин Данила Александрович<br><b>Кафедра:</b> Управления инновациями"},{"id":31000,"title":"Технологический аудит","color":"#0600FF","start":"2023-11-10T09:55:00","end":"2023-11-10T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК110<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Курочкин Данила Александрович<br><b>Кафедра:</b> Управления инновациями"},{"id":26479,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-11-03T09:55:00","end":"2023-11-03T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК110<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":31495,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-12-15T09:55:00","end":"2023-12-15T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК110<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":31496,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-12-01T09:55:00","end":"2023-12-01T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК110<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":31497,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-11-17T09:55:00","end":"2023-11-17T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК110<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32212,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-12-15T13:45:00","end":"2023-12-15T15:15:00","description":"<b>Здание:</b> Бизнес центр<br><b>Аудитория:</b> БЦ217<br><b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32213,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-12-15T11:35:00","end":"2023-12-15T13:05:00","description":"<b>Здание:</b> Бизнес центр<br><b>Аудитория:</b> БЦ217<br><b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32216,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-12-01T13:45:00","end":"2023-12-01T15:15:00","description":"<b>Здание:</b> Бизнес центр<br><b>Аудитория:</b> БЦ217<br><b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32217,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-12-01T11:35:00","end":"2023-12-01T13:05:00","description":"<b>Здание:</b> Бизнес центр<br><b>Аудитория:</b> БЦ217<br><b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32222,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-11-17T13:45:00","end":"2023-11-17T15:15:00","description":"<b>Здание:</b> Бизнес центр<br><b>Аудитория:</b> БЦ217<br><b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32223,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-11-17T11:35:00","end":"2023-11-17T13:05:00","description":"<b>Здание:</b> Бизнес центр<br><b>Аудитория:</b> БЦ217<br><b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32266,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-11-03T13:45:00","end":"2023-11-03T15:15:00","description":"<b>Здание:</b> Бизнес центр<br><b>Аудитория:</b> БЦ217<br><b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32267,"title":"Цифровые технологии управления инновационным проектом","color":"#0600FF","start":"2023-11-03T11:35:00","end":"2023-11-03T13:05:00","description":"<b>Здание:</b> Бизнес центр<br><b>Аудитория:</b> БЦ217<br><b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Гришин Валентин Николаевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32210,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-12-20T09:55:00","end":"2023-12-20T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32211,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-12-20T08:15:00","end":"2023-12-20T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32214,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-12-13T09:55:00","end":"2023-12-13T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32215,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-12-06T09:55:00","end":"2023-12-06T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32218,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-29T09:55:00","end":"2023-11-29T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32220,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-22T09:55:00","end":"2023-11-22T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32224,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-15T09:55:00","end":"2023-11-15T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32226,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-08T09:55:00","end":"2023-11-08T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32268,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-01T09:55:00","end":"2023-11-01T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":32219,"title":"Технологический аудит","color":"#0600FF","start":"2023-11-27T13:45:00","end":"2023-11-27T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Курочкин Данила Александрович<br><b>Кафедра:</b> Управления инновациями"},{"id":32221,"title":"Технологический аудит","color":"#0600FF","start":"2023-11-20T13:45:00","end":"2023-11-20T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Курочкин Данила Александрович<br><b>Кафедра:</b> Управления инновациями"},{"id":32225,"title":"Технологический аудит","color":"#0600FF","start":"2023-11-13T13:45:00","end":"2023-11-13T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Курочкин Данила Александрович<br><b>Кафедра:</b> Управления инновациями"},{"id":32265,"title":"Технологический аудит","color":"#0600FF","start":"2023-11-06T13:45:00","end":"2023-11-06T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Курочкин Данила Александрович<br><b>Кафедра:</b> Управления инновациями"},{"id":25035,"title":"Коммерческая оценка инвестиций","color":"#0600FF","start":"2023-11-11T13:45:00","end":"2023-11-11T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК321<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Онищенко Светлана Ивановна<br><b>Кафедра:</b> Управления инновациями"},{"id":34717,"title":"Коммерческая оценка инвестиций","color":"#0600FF","start":"2023-11-18T13:45:00","end":"2023-11-18T15:15:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК321<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Онищенко Светлана Ивановна<br><b>Кафедра:</b> Управления инновациями"},{"id":36817,"title":"Технический перевод","color":"#0600FF","start":"2023-11-20T08:15:00","end":"2023-11-20T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК 6эт.<br><b>Событие:</b> Лекция<br><b>Кафедра:</b> Иностранных языков"},{"id":36818,"title":"Технический перевод","color":"#0600FF","start":"2023-11-06T08:15:00","end":"2023-11-06T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК 6эт.<br><b>Событие:</b> Лекция<br><b>Кафедра:</b> Иностранных языков"},{"id":39004,"title":"Физическая культура и спорт","color":"#0600FF","start":"2023-12-21T11:35:00","end":"2023-12-21T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39044,"title":"Физическая культура и спорт","color":"#0600FF","start":"2023-12-14T11:35:00","end":"2023-12-14T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39046,"title":"Физическая культура и спорт","color":"#0600FF","start":"2023-12-07T11:35:00","end":"2023-12-07T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39048,"title":"Физическая культура и спорт","color":"#0600FF","start":"2023-11-30T11:35:00","end":"2023-11-30T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39050,"title":"Физическая культура и спорт","color":"#0600FF","start":"2023-11-23T11:35:00","end":"2023-11-23T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39052,"title":"Физическая культура и спорт","color":"#0600FF","start":"2023-11-16T11:35:00","end":"2023-11-16T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39054,"title":"Физическая культура и спорт","color":"#0600FF","start":"2023-11-09T11:35:00","end":"2023-11-09T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39055,"title":"Физическая культура и спорт","color":"#0600FF","start":"2023-11-02T11:35:00","end":"2023-11-02T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39043,"title":"Физическая культура и спорт (элективная дисциплина)","color":"#0600FF","start":"2023-12-16T11:35:00","end":"2023-12-16T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39045,"title":"Физическая культура и спорт (элективная дисциплина)","color":"#0600FF","start":"2023-12-09T11:35:00","end":"2023-12-09T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39047,"title":"Физическая культура и спорт (элективная дисциплина)","color":"#0600FF","start":"2023-12-02T11:35:00","end":"2023-12-02T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39049,"title":"Физическая культура и спорт (элективная дисциплина)","color":"#0600FF","start":"2023-11-25T11:35:00","end":"2023-11-25T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39051,"title":"Физическая культура и спорт (элективная дисциплина)","color":"#0600FF","start":"2023-11-18T11:35:00","end":"2023-11-18T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":39053,"title":"Физическая культура и спорт (элективная дисциплина)","color":"#0600FF","start":"2023-11-11T11:35:00","end":"2023-11-11T13:05:00","description":"<b>Здание:</b> Спортивный комплекс<br><b>Аудитория:</b> ФОК<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Физической культуры"},{"id":47433,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-12-14T08:15:00","end":"2023-12-14T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":47434,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-12-07T08:15:00","end":"2023-12-07T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":47435,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-30T08:15:00","end":"2023-11-30T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":47436,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-23T08:15:00","end":"2023-11-23T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":47437,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-16T08:15:00","end":"2023-11-16T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":47438,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-09T08:15:00","end":"2023-11-09T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":47439,"title":"Разработка стратегии продвижения инновационного продукта","color":"#0600FF","start":"2023-11-02T08:15:00","end":"2023-11-02T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br><b>Кафедра:</b> Управления инновациями"},{"id":47646,"title":"Технический перевод","color":"#0600FF","start":"2023-11-27T09:55:00","end":"2023-11-27T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК 6эт.<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Иностранных языков"},{"id":47647,"title":"Технический перевод","color":"#0600FF","start":"2023-11-20T09:55:00","end":"2023-11-20T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК 6эт.<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Иностранных языков"},{"id":47648,"title":"Технический перевод","color":"#0600FF","start":"2023-11-13T09:55:00","end":"2023-11-13T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК 6эт.<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Иностранных языков"},{"id":47649,"title":"Технический перевод","color":"#0600FF","start":"2023-11-06T09:55:00","end":"2023-11-06T11:25:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК 6эт.<br><b>Событие:</b> Практическое занятие<br><b>Кафедра:</b> Иностранных языков"},{"id":51306,"title":"Цифровая логистика инновационного проекта","color":"#0600FF","start":"2023-12-04T11:35:00","end":"2023-12-04T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК109<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Зубаков Геннадий Викторович<br><b>Кафедра:</b> Логистики"},{"id":51307,"title":"Цифровая логистика инновационного проекта","color":"#0600FF","start":"2023-11-20T11:35:00","end":"2023-11-20T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК109<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Зубаков Геннадий Викторович<br><b>Кафедра:</b> Логистики"},{"id":51308,"title":"Цифровая логистика инновационного проекта","color":"#0600FF","start":"2023-11-06T11:35:00","end":"2023-11-06T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК109<br><b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Зубаков Геннадий Викторович<br><b>Кафедра:</b> Логистики"},{"id":58067,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-12-13T11:35:00","end":"2023-12-13T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":58068,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-12-06T11:35:00","end":"2023-12-06T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":58069,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-29T11:35:00","end":"2023-11-29T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":58070,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-22T11:35:00","end":"2023-11-22T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":58071,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-15T11:35:00","end":"2023-11-15T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":58072,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-08T11:35:00","end":"2023-11-08T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":58073,"title":"Организация управления стартапом","color":"#0600FF","start":"2023-11-01T11:35:00","end":"2023-11-01T13:05:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Дегтярёва Виктория Владимировна<br><b>Кафедра:</b> Управления инновациями"},{"id":21614,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-12-13T08:15:00","end":"2023-12-13T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":21615,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-12-06T08:15:00","end":"2023-12-06T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":21616,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-29T08:15:00","end":"2023-11-29T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":58130,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-22T08:15:00","end":"2023-11-22T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":58131,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-15T08:15:00","end":"2023-11-15T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":58132,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-08T08:15:00","end":"2023-11-08T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":58133,"title":"Финансирование инновационной деятельности","color":"#0600FF","start":"2023-11-01T08:15:00","end":"2023-11-01T09:45:00","description":"<b>Здание:</b> Лабораторный корпус<br><b>Аудитория:</b> ЛК402<br><b>Событие:</b> Лекция<br><b>Преподаватель:</b> Морковкин Дмитрий Евгеньевич<br><b>Кафедра:</b> Управления инновациями"},{"id":158,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Инженерные основы инновационной деятельности","start":"2020-11-07T11:35:00","end":"2020-11-07T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Фаюстов Анатолий Афанасьевич<br>"},{"id":159,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Инженерные основы инновационной деятельности","start":"2020-11-07T13:15:00","end":"2020-11-07T14:45:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Фаюстов Анатолий Афанасьевич<br>"},{"id":160,"color":"#f57900","online":1,"title":"ОНЛАЙН [ЛР] Инженерные основы инновационной деятельности","start":"2020-11-07T15:25:00","end":"2020-11-07T16:55:00","description":"<b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Фаюстов Анатолий Афанасьевич<br>"},{"id":345,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] История (история России, всеобщая история)","start":"2020-11-03T15:25:00","end":"2020-11-03T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Ильюхов Александр Антонович<br>"},{"id":557,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Математика","start":"2020-11-11T13:15:00","end":"2020-11-11T14:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Бухтоярова Марина Васильевна<br>"},{"id":774,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] История (история России, всеобщая история)","start":"2020-11-10T15:25:00","end":"2020-11-10T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Ильюхов Александр Антонович<br>"},{"id":983,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Инженерные основы инновационной деятельности","start":"2020-11-14T11:35:00","end":"2020-11-14T16:55:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Фаюстов Анатолий Афанасьевич<br>"},{"id":1179,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Безопасность жизнедеятельности","start":"2020-11-13T13:15:00","end":"2020-11-13T14:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Аракелова Галина Александровна<br>"},{"id":1351,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Основы профессионального развития","start":"2020-11-13T11:35:00","end":"2020-11-13T13:05:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Алпатова Екатерина Александровна<br>"},{"id":1353,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Основы профессионального развития","start":"2020-11-13T15:25:00","end":"2020-11-13T16:55:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Алпатова Екатерина Александровна<br>"},{"id":1923,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Логика и критическое мышление","start":"2020-11-09T11:35:00","end":"2020-11-09T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Яценко Оксана Юрьевна<br>"},{"id":2074,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Безопасность жизнедеятельности","start":"2020-11-16T15:25:00","end":"2020-11-16T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Чернышенко Сергей Викторович<br>"},{"id":2538,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] История (история России, всеобщая история)","start":"2020-11-16T11:35:00","end":"2020-11-16T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Надехина Юлия Петровна<br>"},{"id":2872,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Математика","start":"2020-11-18T11:35:00","end":"2020-11-18T13:05:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Бухтоярова Марина Васильевна<br>"},{"id":2880,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Математика","start":"2020-11-18T13:15:00","end":"2020-11-18T14:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Бухтоярова Марина Васильевна<br>"},{"id":2933,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] История (история России, всеобщая история)","start":"2020-11-17T15:25:00","end":"2020-11-17T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Ильюхов Александр Антонович<br>"},{"id":3333,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Философия","start":"2020-11-18T15:25:00","end":"2020-11-18T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Захаров Михаил Юрьевич<br>"},{"id":3481,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Безопасность жизнедеятельности","start":"2020-11-20T13:15:00","end":"2020-11-20T14:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Аракелова Галина Александровна<br>"},{"id":3533,"color":"#f57900","online":1,"title":"ОНЛАЙН [ЛР] Инженерные основы инновационной деятельности","start":"2020-11-21T13:15:00","end":"2020-11-21T14:45:00","description":"<b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Фаюстов Анатолий Афанасьевич<br>"},{"id":3619,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Информационные системы и технологии","start":"2020-11-20T17:05:00","end":"2020-11-20T18:35:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Белоусов Виталий Андреевич<br>"},{"id":3761,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Основы профессионального развития","start":"2020-11-20T15:25:00","end":"2020-11-20T16:55:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Алпатова Екатерина Александровна<br>"},{"id":3922,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Инженерные основы инновационной деятельности","start":"2020-11-21T11:35:00","end":"2020-11-21T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Фаюстов Анатолий Афанасьевич<br>"},{"id":3923,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Инженерные основы инновационной деятельности","start":"2020-11-21T15:25:00","end":"2020-11-21T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Фаюстов Анатолий Афанасьевич<br>"},{"id":4020,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Математика","start":"2020-11-25T13:15:00","end":"2020-11-25T14:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Бухтоярова Марина Васильевна<br>"},{"id":4570,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] История (история России, всеобщая история)","start":"2020-11-30T11:35:00","end":"2020-11-30T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Надехина Юлия Петровна<br>"},{"id":4614,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Логика и критическое мышление","start":"2020-11-23T11:35:00","end":"2020-11-23T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Яценко Оксана Юрьевна<br>"},{"id":4933,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] История (история России, всеобщая история)","start":"2020-11-24T15:25:00","end":"2020-11-24T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Ильюхов Александр Антонович<br>"},{"id":5235,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Философия","start":"2020-11-25T11:35:00","end":"2020-11-25T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Захаров Михаил Юрьевич<br>"},{"id":5237,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Философия","start":"2020-11-25T15:25:00","end":"2020-11-25T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Захаров Михаил Юрьевич<br>"},{"id":5723,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Инженерные основы инновационной деятельности","start":"2020-11-28T11:35:00","end":"2020-11-28T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Фаюстов Анатолий Афанасьевич<br>"},{"id":6375,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Математика","start":"2020-12-02T11:35:00","end":"2020-12-02T13:05:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Бухтоярова Марина Васильевна<br>"},{"id":6561,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Философия","start":"2020-12-02T15:25:00","end":"2020-12-02T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Захаров Михаил Юрьевич<br>"},{"id":7285,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Логика и критическое мышление","start":"2020-12-07T11:35:00","end":"2020-12-07T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Яценко Оксана Юрьевна<br>"},{"id":7337,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] История (история России, всеобщая история)","start":"2020-12-14T11:35:00","end":"2020-12-14T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Надехина Юлия Петровна<br>"},{"id":7584,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Философия","start":"2020-12-09T11:35:00","end":"2020-12-09T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Захаров Михаил Юрьевич<br>"},{"id":7586,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Философия","start":"2020-12-09T15:25:00","end":"2020-12-09T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Захаров Михаил Юрьевич<br>"},{"id":8465,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Математика","start":"2020-12-16T00:00:00","end":"2020-12-16T13:05:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Бухтоярова Марина Васильевна<br>"},{"id":8542,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Философия","start":"2020-12-16T15:25:00","end":"2020-12-16T16:55:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Захаров Михаил Юрьевич<br>"},{"id":9065,"color":"#edd400","online":1,"title":"ОНЛАЙН [Зачет] Математика","start":"2020-12-23T09:00:00","end":"2020-12-23T10:30:00","description":"<b>Событие:</b> Зачет<br><b>Преподаватель:</b> Бухтоярова Марина Васильевна<br>"},{"id":18732,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Командная работа и лидерство","start":"2021-11-01T09:55:00","end":"2021-11-01T11:25:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Сувалова Татьяна Викторовна<br>"},{"id":20102,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Управление инновационной деятельностью","start":"2021-11-02T08:15:00","end":"2021-11-02T09:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Камчатова Екатерина Юрьевна<br>"},{"id":20103,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Управление инновационной деятельностью","start":"2021-11-02T09:55:00","end":"2021-11-02T11:25:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Камчатова Екатерина Юрьевна<br>"},{"id":20310,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Статистический анализ в бизнесе","start":"2021-11-03T08:15:00","end":"2021-11-03T09:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Теплякова Марина Юрьевна<br>"},{"id":20311,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Статистический анализ в бизнесе","start":"2021-11-03T09:55:00","end":"2021-11-03T11:25:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Теплякова Марина Юрьевна<br>"},{"id":20579,"color":"#f57900","online":0,"title":" [ПЗ] Теория систем управления и принятия решений","start":"2021-11-02T13:45:00","end":"2021-11-02T15:15:00","description":"<b>Событие:</b> Практическое занятие<br><b>Аудитория:</b> ЛК101<br><b>Преподаватель:</b> Тимофеева Татьяна Борисовна<br>"},{"id":20749,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Правоведение","start":"2021-11-09T08:15:00","end":"2021-11-09T09:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Целуйко Андрей Валерьевич<br>"},{"id":21033,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Командная работа и лидерство","start":"2021-11-08T08:15:00","end":"2021-11-08T09:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Сувалова Татьяна Викторовна<br>"},{"id":21034,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Командная работа и лидерство","start":"2021-11-08T09:55:00","end":"2021-11-08T11:25:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Сувалова Татьяна Викторовна<br>"},{"id":21461,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Теория систем управления и принятия решений","start":"2021-11-09T11:35:00","end":"2021-11-09T13:05:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Тимофеева Татьяна Борисовна<br>"},{"id":21462,"color":"#f57900","online":0,"title":" [ПЗ] Теория систем управления и принятия решений","start":"2021-11-09T13:45:00","end":"2021-11-09T15:15:00","description":"<b>Событие:</b> Практическое занятие<br><b>Аудитория:</b> ЛК101<br><b>Преподаватель:</b> Тимофеева Татьяна Борисовна<br>"},{"id":21505,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Управление инновационной деятельностью","start":"2021-11-09T09:55:00","end":"2021-11-09T11:25:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Камчатова Екатерина Юрьевна<br>"},{"id":22032,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Командная работа и лидерство","start":"2021-11-15T09:55:00","end":"2021-11-15T11:25:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Сувалова Татьяна Викторовна<br>"},{"id":22675,"color":"#f57900","online":1,"title":"ОНЛАЙН [Л] Командная работа и лидерство","start":"2021-11-22T08:15:00","end":"2021-11-22T09:45:00","description":"<b>Событие:</b> Лекция<br><b>Преподаватель:</b> Сувалова Татьяна Викторовна<br>"},{"id":22676,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Командная работа и лидерство","start":"2021-11-22T09:55:00","end":"2021-11-22T11:25:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Сувалова Татьяна Викторовна<br>"},{"id":23218,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Командная работа и лидерство","start":"2021-11-29T09:55:00","end":"2021-11-29T11:25:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Сувалова Татьяна Викторовна<br>"},{"id":24216,"color":"#edd400","online":1,"title":"ОНЛАЙН [Зачет] Командная работа и лидерство","start":"2021-12-27T09:00:00","end":"2021-12-27T11:00:00","description":"<b>Событие:</b> Зачет<br><b>Преподаватель:</b> Сувалова Татьяна Викторовна<br>"},{"id":26329,"color":"#f57900","online":1,"title":"ОНЛАЙН [ПЗ] Проектная работа 1","start":"2022-11-08T11:35:00","end":"2022-11-08T13:05:00","description":"<b>Событие:</b> Практическое занятие<br><b>Преподаватель:</b> Онищенко Светлана Ивановна<br>"},{"id":26586,"color":"#f57900","online":1,"title":"ОНЛАЙН [ЛР] Оценка эффективности программы продвижения","start":"2022-12-30T11:00:00","end":"2022-12-30T12:00:00","description":"<b>Событие:</b> Лабораторная работа<br><b>Преподаватель:</b> Береговская Татьяна Александровна<br>"}]
